+++ src/main/java/com/buuz135/industrial/proxy/ConveyorRegistry.java
@@ -1,9 +1,9 @@
 package com.buuz135.industrial.proxy;
 
 import com.buuz135.industrial.api.conveyor.ConveyorUpgradeFactory;
-import com.buuz135.industrial.proxy.block.ConveyorDetectorUpgrade;
-import com.buuz135.industrial.proxy.block.ConveyorExtractionUpgrade;
-import com.buuz135.industrial.proxy.block.ConveyorInsertionUpgrade;
+import com.buuz135.industrial.proxy.block.upgrade.ConveyorDetectorUpgrade;
+import com.buuz135.industrial.proxy.block.upgrade.ConveyorExtractionUpgrade;
+import com.buuz135.industrial.proxy.block.upgrade.ConveyorInsertionUpgrade;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.registries.IForgeRegistry;
